Deprecate Location.detect
We're only showing the deprecation warning for locations other
than 'auto', since auto already have it's own deprecation message.

We're also moving the rootURL trailing slash check from detect
to initState. That assert has nothing to do with detection, but it
was likely added there because that method is called *after*
the rootURL value is set on the location (by the router).

Luckily, the initState method is also called after that value is set,
so we can use that hook instead.

We're hijacking a method for something that isn't it's primary purpose,
but that was the case with the existing code too, so this seems
like a reasonable solution.